PRSD-495: Moved getManualAddress to JourneyDataHelper
