+++ src/main/kotlin/uk/gov/communities/prsdb/webapp/services/AddressDataService.kt
@@ -7,7 +7,6 @@ import org.springframework.context.annotation.Scope
 import org.springframework.context.annotation.ScopedProxyMode
 import org.springframework.stereotype.Service
 import org.springframework.web.context.WebApplicationContext
-import uk.gov.communities.prsdb.webapp.forms.journeys.JourneyData
 import uk.gov.communities.prsdb.webapp.forms.journeys.objectToStringKeyedMap
 import uk.gov.communities.prsdb.webapp.models.dataModels.AddressDataModel
 
@@ -42,54 +41,5 @@ class AddressDataService(
         cachedResults[uprn.toString()] = result
 
         session.setAttribute("addressRegisteredResults", cachedResults)
-    }
-
-    fun getManualAddress(
-        journeyDataService: JourneyDataService,
-        journeyData: JourneyData,
-        manualAddressPathSegment: String,
-    ): AddressDataModel? {
-        val addressLineOne =
-            journeyDataService.getFieldStringValue(
-                journeyData,
-                manualAddressPathSegment,
-                "addressLineOne",
-            ) ?: return null
-
-        val townOrCity =
-            journeyDataService.getFieldStringValue(
-                journeyData,
-                manualAddressPathSegment,
-                "townOrCity",
-            ) ?: return null
-
-        val postcode =
-            journeyDataService.getFieldStringValue(
-                journeyData,
-                manualAddressPathSegment,
-                "postcode",
-            ) ?: return null
-
-        val addressLineTwo =
-            journeyDataService.getFieldStringValue(
-                journeyData,
-                manualAddressPathSegment,
-                "addressLineTwo",
-            )
-
-        val county =
-            journeyDataService.getFieldStringValue(
-                journeyData,
-                manualAddressPathSegment,
-                "county",
-            )
-
-        return AddressDataModel.fromManualAddressData(
-            addressLineOne,
-            townOrCity,
-            postcode,
-            addressLineTwo,
-            county,
-        )
     }
 }
